drop module shim
