+++ package.json
@@ -6,8 +6,7 @@
   "scripts": {
     "build:remote": "node build-remote.js",
     "build:host": "node build-host.js",
-    "build:shim": "node build-shim.js",
-    "build": "npm run build:shim && npm run build:host && npm run build:remote",
+    "build": "npm run build:host && npm run build:remote",
     "start": "serve dist",
     "dev": "npm run build && npm run start"
   },
